refactor: preapre to add more languages support
