+++ lua/template-string/langs/init.lua
@@ -1,0 +1,3 @@
+return {
+  js = require("template-string.langs.javascript"),
+}
