--- lua/template-string/quote_string.lua
@@ -1,36 +1,0 @@
-local U = require("template-string.utils")
-local C = require("template-string.config")
-local M = {}
-
-function M.is_already_template(text)
-	return text:match("^{?`.*`}?$") ~= nil
-end
-
-function M.handle_quote_string(node, buf)
-	local text = vim.treesitter.get_node_text(node, buf)
-
-	if
-		U.is_undo_or_redo()
-		or not U.has_template_string(text)
-		-- ignore last redo
-		or M.is_already_template(text)
-		-- ignore multiline quote string (treesitter could detect quote string in bad syntax)
-		or U.is_multiline(text)
-	then
-		return
-	end
-
-	local new_text = U.replace_quotes(text, U.quote.BACKTICK)
-
-	-- add brackets if it's jsx attribute
-	if C.options.jsx_brackets and U.is_jsx_node(node) then
-		new_text = "{" .. new_text .. "}"
-
-		-- move the cursor to stay in the same position after adding brackets
-		U.move_cursor({ 0, 1 })
-	end
-
-	U.replace_node_text(node, buf, new_text)
-end
-
-return M
